requirements.txt
Click==7.0
PyMySQL==0.9.3
PyYAML==5.2